requirements.txt
aiohttp==3.7.4
async_sender==1.4.3
facet==0.7.0
loguru==0.5.3
pydantic[dotenv]==1.8.2
tortoise-orm[asyncpg]==0.17.2
yarl==1.6.3
